pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "attendance-sheet-generator"
version = "0.1.0"
description = "通用培训考勤表自动生成工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "python-docx>=0.8.11",
    "pdfplumber>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyinstaller>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
